feature/MVC-file-upload-controller
To load the Spring MVC file, we use our MultipartResolver interface (we will use the CommonsMultipartResolver class that implements it). Work with him is very simple. All you need to do is specify the mapping by which the controller intercepts the request to download the file. Further in the class, standard work with byte streams and file recording are used.

added

    - FileUploadController.java
    descriptions

added

    -Dockerfile (test VM)
    -Jenkinsfile changes